refactor : PaymentClient 인터페이스 추출, HttpHeader 선언부 수정
